# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cube-back",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ressources": {
            "get": {
                "tags": [
                    "Ressource"
                ],
                "summary": "Récupérer toutes les ressources",
                "operationId": "9e2c3350f2d7a4aa04c9296e672dd45a",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items to return per page (1-100)",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort items by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "keywordTitreRes",
                        "in": "query",
                        "description": "Keyword for searching items by keywordTitreRes",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keywordIdTypeRes",
                        "in": "query",
                        "description": "Keyword for searching items by keywordIdTypeRes",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keywordIdRel",
                        "in": "query",
                        "description": "Keyword for searching items by keywordIdRel",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keywordIdVis",
                        "in": "query",
                        "description": "Keyword for searching items by keywordIdVis",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keywordIdCat",
                        "in": "query",
                        "description": "Keyword for searching items by keywordIdCat",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keywordIdCreateur",
                        "in": "query",
                        "description": "Keyword for searching items by keywordIdCreateur",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ressource"
                ],
                "summary": "Créer une ressource",
                "operationId": "d5e8c5df1d1cac7ff2c1a25f81ffb388",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre_res",
                                    "contenu_res",
                                    "id_type_res",
                                    "id_vis",
                                    "id_createur",
                                    "arrayIdCat",
                                    "arrayIdRel"
                                ],
                                "properties": {
                                    "titre_res": {
                                        "type": "string",
                                        "maxLength": 40
                                    },
                                    "contenu_res": {
                                        "type": "string"
                                    },
                                    "url_res": {
                                        "type": "string"
                                    },
                                    "id_type_res": {
                                        "type": "integer"
                                    },
                                    "id_vis": {
                                        "type": "integer"
                                    },
                                    "id_createur": {
                                        "type": "integer"
                                    },
                                    "arrayIdRel": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "arrayIdCat": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/ressources/{id}": {
            "get": {
                "tags": [
                    "Ressource"
                ],
                "summary": "Récupérer une ressource spécifique",
                "operationId": "70fa20537068ba16ca397172f7923e17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ressource"
                ],
                "summary": "Modifier une ressource",
                "operationId": "ac433b3f23904630aaf553ab263a3e27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre_res",
                                    "contenu_res",
                                    "id_type_res",
                                    "id_vis",
                                    "id_createur",
                                    "arrayIdCat",
                                    "arrayIdRel",
                                    "is_archive"
                                ],
                                "properties": {
                                    "titre_res": {
                                        "type": "string",
                                        "maxLength": 40
                                    },
                                    "contenu_res": {
                                        "type": "string"
                                    },
                                    "url_res": {
                                        "type": "string"
                                    },
                                    "id_type_res": {
                                        "type": "integer"
                                    },
                                    "id_vis": {
                                        "type": "integer"
                                    },
                                    "id_createur": {
                                        "type": "integer"
                                    },
                                    "arrayIdRel": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "arrayIdCat": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "is_archive": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item updated successfully"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ressource"
                ],
                "summary": "Supprimer une ressource",
                "operationId": "b1b9d669f5d1361f8378633eb52d0bdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item deleted successfully"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}